fix(arithmetic): use i128 intermediate for millionths rate calculations to prevent overflow
Four rate/ratio calculations that multiplied a counter by MILLION (1,000,000)
before dividing would silently overflow i64 for large inputs. Upgrade each
to use i128 for the intermediate product:

- PolicyComparisonReport::divergence_rate_millionths
- FailoverController::failover_success_rate_millionths
- QueueArtifact::completion_rate_millionths
- Novelty entropy bonus in synthesize_candidate (also remove redundant
  .min(200_000) clamp since % 200_001 modulus already bounds the result)

+++ crates/franken-engine/src/counterfactual_replay_engine.rs
@@ -247,7 +247,7 @@ impl PolicyComparisonReport {
         if self.decisions_evaluated == 0 {
             return 0;
         }
-        (self.divergence_count as i64 * MILLION) / self.decisions_evaluated as i64
+        (self.divergence_count as i128 * MILLION as i128 / self.decisions_evaluated as i128) as i64
     }
 }
